// src/types/enum/ClientStatusCode.ts
export enum ClientStatusCode {
  BadRequest = 400,
  Conflict = 409,
  ExpectationFailed = 417,
  FailedDependency = 424,
  Forbidden = 403,
  Gone = 410,
  ImATeapot = 418,
  LengthRequired = 411,
  Locked = 423,
  LoginTimeOut = 440,
  MethodNotAllowed = 405,
  MisdirectedRequest = 421,
  NotAcceptable = 406,
  NotFound = 404,
  PayloadTooLarge = 413,
  PaymentRequired = 402,
  PreconditionFailed = 412,
  PreconditionRequired = 428,
  ProxyAuthRequired = 407,
  RangeNotSatisfiable = 416,
  RequestHeaderFieldsTooLarge = 431,
  RequestTimeout = 408,
  RetryWith = 449,
  TooManyRequests = 429,
  Unauthorized = 401,
  UnavailableForLegalReasons = 451,
  UnprocessableEntity = 422,
  UnsupportedMediaType = 415,
  UpgradeRequired = 426,
  URITooLong = 414,
}